removed old code, added ludwig db.predictor updates
