+++ mindsdb/integrations/handlers/lightwood_handler/lightwood_handler/learn_process.py
@@ -22,6 +22,7 @@ from mindsdb.utilities.config import Config
 from mindsdb.utilities.functions import mark_process
 from mindsdb.utilities.log import log
 from mindsdb.integrations.libs.const import PREDICTOR_STATUS
+from mindsdb.integrations.utilities.utils import format_exception_error
 
 
 ctx = mp.get_context('spawn')
@@ -163,17 +164,6 @@ def run_learn_remote(df: DataFrame, predictor_id: int) -> None:
         predictor_record.data['error'] = str(resp.text)
 
     session.commit()
-
-
-def format_exception_error(exception):
-    try:
-        exception_type, _exception_object, exception_traceback = sys.exc_info()
-        filename = exception_traceback.tb_frame.f_code.co_filename
-        line_number = exception_traceback.tb_lineno
-        error_message = f'{exception_type.__name__}: {exception}, raised at: {filename}#{line_number}'
-    except Exception:
-        error_message = str(exception)
-    return error_message
 
 
 @mark_process(name='learn')
